refactored OOP Ruby school library
