+++ student.rb
@@ -3,7 +3,7 @@ require_relative 'person'
 class Student < Person
   attr_reader :classroom, :type
 
-  def initialize(age, classroom, name, parent_permission: true)
+  def initialize(age, name, classroom, parent_permission: true)
     super(age, name, parent_permission: parent_permission)
     @classroom = classroom
     @type = 'student'
